requirements.txt
streamlit
transformers
google-generativeai
python-dotenv
torch
torchvision
torchaudio